Restructure file system
Clear up server.js and use a modular system.
