+++ routes/user.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+const { register } = require('../controllers/register');
+
+router.post('/register', register);
+
+module.exports = router;
