+++ utils/hash.js
@@ -1,0 +1,17 @@
+const crypto = require('crypto');
+
+/**
+ * Hashes text with the sha256 algorithm into base64 encoding
+ * 
+ * @param {string} input -  text to be hashed
+ * @return {string} base64 encoded hash
+ */
+const shaHash = (input) =>
+{
+    // input: text to be hashed
+    // output: hashed text (base64 encoded)
+
+    return crypto.createHash('sha256').update(input).digest('base64');
+};
+
+module.exports = { shaHash };
